+++ packages/components/src/components.d.ts
@@ -306,6 +306,10 @@ export namespace Components {
          */
         "iconBadgeValue": string;
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level": string;
+        /**
           * @description Location object.
          */
         "location": any;
@@ -873,6 +877,10 @@ export namespace Components {
           * Id for the input field.
          */
         "idAttribute": string;
+        /**
+          * The language used when retrieving Google Places or Mapbox autocomplete suggestions.
+         */
+        "language": string;
         /**
           * If searching should include Mapbox autocomplete suggestions.
          */
@@ -1665,6 +1673,10 @@ declare namespace LocalJSX {
          */
         "iconBadgeValue"?: string;
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level"?: string;
+        /**
           * @description Location object.
          */
         "location"?: any;
@@ -2204,6 +2216,10 @@ declare namespace LocalJSX {
           * Id for the input field.
          */
         "idAttribute"?: string;
+        /**
+          * The language used when retrieving Google Places or Mapbox autocomplete suggestions.
+         */
+        "language"?: string;
         /**
           * If searching should include Mapbox autocomplete suggestions.
          */