+++ packages/map-template/CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.26.0] - 2023-12-20
+
+### Added
+
+- Add support for French, German and Danish languages. The browser language is used to pick one of those, and it can also be forced with a new `language` property.
+
 ## [1.25.0] - 2023-12-18
 
 ### Added
 
-- Added support for having QR code button when getting directions in Kiosk mode. 
+- Added support for having QR code button when getting directions in Kiosk mode.
 
 ## [1.24.1] - 2023-12-14